elk/api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
psycopg[binary]>=3.1
gunicorn>=21.2,<22.0
